+++ server/src/models/DBconnection.ts
@@ -1,0 +1,22 @@
+const postgres = require('postgres');
+import dotenv from 'dotenv';
+dotenv.config();
+
+let { PGHOST, PGDATABASE, PGUSER, PGPASSWORD, ENDPOINT_ID } = process.env;
+
+const sql = postgres({
+  host: PGHOST,
+  database: PGDATABASE,
+  username: PGUSER,
+  password: PGPASSWORD,
+  port: 5432,
+  ssl: 'require',
+  connection: {
+    options: `project=${ENDPOINT_ID}`,
+  },
+});
+
+export async function getPgVersion() {
+  const result = await sql`select version()`;
+  console.log(result[0]);
+}
